Extract `get_model_class`, `get_loaded_model`, `get_loaded_model_by_path`
