+++ lib/annotate_rb/model_annotator.rb
@@ -12,5 +12,6 @@ module AnnotateRb
     autoload :FileNameResolver, 'annotate_rb/model_annotator/file_name_resolver'
     autoload :FileAnnotationRemover, 'annotate_rb/model_annotator/file_annotation_remover'
     autoload :AnnotationPatternGenerator, 'annotate_rb/model_annotator/annotation_pattern_generator'
+    autoload :ModelClassGetter, 'annotate_rb/model_annotator/model_class_getter'
   end
 end
